requirements.txt
aspose-cells==23.12.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
findspark==2.0.1
Flask==3.0.0
gunicorn==20.0.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.5.0
MarkupSafe==2.1.3
multi==0.0.1
multiprocess==0.70.15
mysql-connector==2.2.9
numpy==1.26.3
packaging==23.2
pandas==2.1.4
py4j==0.10.9.7
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
wheel==0.42.0
